requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
zeep==4.2.1
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
click==8.1.7
schedule==1.2.0